% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{dbConnection}
\alias{dbConnection}
\title{dbConnection class}
\description{
connect to database using either DBI or pool
}
\details{
dbConnection class

connect using DBI or pool
}
\section{Methods}{

\describe{
\item{\strong{connect}}{connect to database}
\item{\strong{close}}{close connection to database}
\item{\strong{dbSendQuery}}{send query (statement) to connection}
\item{\strong{dbGetQuery}}{send query to connection}
}
}

\examples{
dbConnection$new() # creates new object

dbConnection$connect(
  drv = RSQLite::SQLite(),
  dbname = "mydatabase.sqlite",
  usepool = FALSE
)
# sets $DBIconn or $poolcon to connection
# both set to NULL if fail/warn
# by default 'usepool' is TRUE and attempts
# to use 'pool' package (if available)
# if 'usepool' is FALSE, then do not use
# the 'pool' package, so use 'DBI' instead

dbConnection$close() # closes all connections

query <- "UPDATE Users SET Name = ? WHERE id = ?"
data_for_sql <- list(c("John"), id)
dbConnection$dbSendQuery(query, data_for_sql)
# send parametized 'SQL query'
# with 'data_for_sql'
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{DBIconn}}{- connection using DBI (NULL if not connected)}

\item{\code{poolconn}}{- connection using pool (NULL is not connected)}

\item{\code{conn()}}{- connection using whichever connection is available}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{conn()}}{- connection using whichever connection is available}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-connect}{\code{dbConnection$connect()}}
\item \href{#method-close}{\code{dbConnection$close()}}
\item \href{#method-is_open}{\code{dbConnection$is_open()}}
\item \href{#method-finalize}{\code{dbConnection$finalize()}}
\item \href{#method-conn}{\code{dbConnection$conn()}}
\item \href{#method-dbSendQuery}{\code{dbConnection$dbSendQuery()}}
\item \href{#method-dbSendGlueQuery}{\code{dbConnection$dbSendGlueQuery()}}
\item \href{#method-dbGetQuery}{\code{dbConnection$dbGetQuery()}}
\item \href{#method-open_log_db}{\code{dbConnection$open_log_db()}}
\item \href{#method-write_log_db}{\code{dbConnection$write_log_db()}}
\item \href{#method-duration_log_db}{\code{dbConnection$duration_log_db()}}
\item \href{#method-close_log_db}{\code{dbConnection$close_log_db()}}
\item \href{#method-clone}{\code{dbConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$connect(drv, ..., usepool = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_open"></a>}}
\if{latex}{\out{\hypertarget{method-is_open}{}}}
\subsection{Method \code{is_open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$is_open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-conn"></a>}}
\if{latex}{\out{\hypertarget{method-conn}{}}}
\subsection{Method \code{conn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$conn(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dbSendQuery"></a>}}
\if{latex}{\out{\hypertarget{method-dbSendQuery}{}}}
\subsection{Method \code{dbSendQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$dbSendQuery(query, data_for_sql)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dbSendGlueQuery"></a>}}
\if{latex}{\out{\hypertarget{method-dbSendGlueQuery}{}}}
\subsection{Method \code{dbSendGlueQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$dbSendGlueQuery(query, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dbGetQuery"></a>}}
\if{latex}{\out{\hypertarget{method-dbGetQuery}{}}}
\subsection{Method \code{dbGetQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$dbGetQuery(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open_log_db"></a>}}
\if{latex}{\out{\hypertarget{method-open_log_db}{}}}
\subsection{Method \code{open_log_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$open_log_db(filename, tag)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_log_db"></a>}}
\if{latex}{\out{\hypertarget{method-write_log_db}{}}}
\subsection{Method \code{write_log_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$write_log_db(query = "", data = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-duration_log_db"></a>}}
\if{latex}{\out{\hypertarget{method-duration_log_db}{}}}
\subsection{Method \code{duration_log_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$duration_log_db(log_detail)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close_log_db"></a>}}
\if{latex}{\out{\hypertarget{method-close_log_db}{}}}
\subsection{Method \code{close_log_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$close_log_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dbConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
